+++ backend/resolvers/badgeResolver.ts
@@ -4,40 +4,82 @@ import {
   Icon,
   Badge,
   BadgeLevel,
+  PrismaClient,
 } from "@prisma/client";
-import BadgeService from "../services/implementation/badgeImplementation";
-import IBadgeService from "../services/interface/badgeInterface";
 
-const badgeService: IBadgeService = new BadgeService();
+import { getNow } from "../utils/formatDateTime";
+
+const prisma = new PrismaClient();
 
 const badgeResolver = {
   Query: {
     getCustomBadges: async (): Promise<Badge[]> => {
-      return badgeService.getCustomBadges();
+      return prisma.badge.findMany({
+        where: {
+          badge_type: "CUSTOM",
+        },
+        include: {
+          badge_level: true,
+        },
+        orderBy: {
+          name: "asc",
+        },
+      });
     },
     getSystemBadges: async (): Promise<Badge[]> => {
-      return badgeService.getSystemBadges();
-    },
-    getEarnedBadgesByParticipant: async (
-        _parent: undefined,
-        { participantId }: { participantId: number }
-        ): Promise<EarnedBadge[]> => {
-            return badgeService.getEarnedBadgesByParticipant(participantId);
+      return prisma.badge.findMany({
+        where: {
+          badge_type: "SYSTEM",
         },
+        include: {
+          badge_level: {
+            orderBy: {
+              level: "asc",
+            },
+          },
+        },
+        orderBy: {
+          name: "asc",
+        },
+      });
+    },
   },
+  getEarnedBadgesByParticipant: async (
+    _parent: undefined,
+    { participantId }: { participantId: number }
+  ): Promise<EarnedBadge[]> => {
+    return prisma.earnedBadge.findMany({
+      where: { participant_id: participantId },
+      orderBy: { date_received: "desc" },
+    });
+  },
+
   Mutation: {
     updateBadgeStatus: async (
       _parent: undefined,
       {
         badge_id,
-        is_active
+        is_active,
       }: {
-        badge_id: number,
-        is_active: boolean,
+        badge_id: number;
+        is_active: boolean;
       }
     ): Promise<boolean> => {
-      return badgeService.updateBadgeStatus(badge_id, is_active)
+      const badge = await prisma.badge.findUnique({
+        where: { badge_id },
+      });
+      if (!badge) {
+        throw new Error("Badge not found");
+      }
+      await prisma.badge.update({
+        where: { badge_id },
+        data: {
+          is_active,
+        },
+      });
+      return true;
     },
+    // fix: need to also add marillac bucks for participant here
     assignCustomBadge: async (
       _parent: undefined,
       {
@@ -50,12 +92,72 @@ const badgeResolver = {
         participant_ids: number[];
       }
     ): Promise<number[]> => {
-      return badgeService.assignCustomBadge(
-        badge_id,
-        marillac_bucks,
-        participant_ids
+      const badge = await prisma.badge.findUnique({
+        where: { badge_id },
+      });
+      if (!badge) throw new Error(`Badge with ID ${badge_id} not found`);
+
+      const participants = await prisma.participant.findMany({
+        where: {
+          participant_id: {
+            in: participant_ids,
+          },
+        },
+      });
+      if (participants.length !== participant_ids.length)
+        throw new Error(`Some IDs were invalid!`);
+
+      const existingBadges = await prisma.earnedBadge.findMany({
+        where: {
+          participant_id: { in: participant_ids },
+          name: badge.name,
+        },
+        select: { participant_id: true },
+      });
+      const alreadyEarnedIds = new Set(
+        existingBadges.map((b) => b.participant_id)
       );
+      const eligibleParticipantIds = participant_ids.filter(
+        (id) => !alreadyEarnedIds.has(id)
+      );
+
+      if (eligibleParticipantIds.length === 0) {
+        throw new Error("Participants have already received this badge");
+      }
+
+      await prisma.$transaction(
+        eligibleParticipantIds.map((id) =>
+          prisma.earnedBadge.create({
+            data: {
+              badge_id: badge.badge_id,
+              participant_id: id,
+              date_received: new Date().toISOString(),
+              name: badge.name,
+              description: badge.description,
+              badge_icon: badge.icon,
+              level: 1,
+            },
+          })
+        )
+      );
+
+      // also add marillac bucks for the earned badge
+      await prisma.$transaction(
+        eligibleParticipantIds.map((id) =>
+          prisma.participant.update({
+            where: { participant_id: id },
+            data: {
+              marillac_bucks: {
+                increment: marillac_bucks,
+              },
+            },
+          })
+        )
+      );
+
+      return eligibleParticipantIds;
     },
+
     editCustomBadge: async (
       _parent: undefined,
       {
@@ -68,11 +170,24 @@ const badgeResolver = {
         new_custom_badge_description?: string;
       }
     ): Promise<boolean> => {
-      return badgeService.editCustomBadge(
-        custom_badge_id,
-        new_custom_badge_name,
-        new_custom_badge_description
-      );
+      if (
+        new_custom_badge_name == undefined &&
+        new_custom_badge_description == undefined
+      ) {
+        throw new Error("No edits provided");
+      }
+      await prisma.badge.update({
+        where: { badge_id: custom_badge_id, badge_type: "CUSTOM" },
+        data: {
+          ...(new_custom_badge_name !== undefined && {
+            name: new_custom_badge_name,
+          }),
+          ...(new_custom_badge_description !== undefined && {
+            description: new_custom_badge_description,
+          }),
+        },
+      });
+      return true;
     },
     editSystemBadge: async (
       _parent: undefined,
@@ -86,12 +201,24 @@ const badgeResolver = {
         system_badge_criteria?: string;
       }
     ): Promise<boolean> => {
-      return badgeService.editSystemBadge(
-        system_badge_id,
-        system_badge_name,
-        system_badge_criteria
-      );
+      const badge = await prisma.badge.findUnique({
+        where: { badge_id: system_badge_id },
+      });
+      if (!badge || badge.badge_type !== "SYSTEM") {
+        throw new Error("BAdge not found or not a system badge");
+      }
+      await prisma.badge.update({
+        where: { badge_id: system_badge_id },
+        data: {
+          name: system_badge_name,
+          ...(system_badge_criteria !== undefined && {
+            description: system_badge_criteria,
+          }),
+        },
+      });
+      return true;
     },
+
     createCustomBadge: async (
       _parent: undefined,
       {
@@ -104,11 +231,16 @@ const badgeResolver = {
         icon: Icon;
       }
     ): Promise<boolean> => {
-      return badgeService.createCustomBadge(
-        name,
-        description,
-        icon
-      );
+      await prisma.badge.create({
+        data: {
+          name,
+          description,
+          icon,
+          is_consecutive: false,
+          badge_type: "CUSTOM",
+        },
+      });
+      return true;
     },
     deleteCustomBadge: async (
       _parent: undefined,
@@ -118,7 +250,25 @@ const badgeResolver = {
         badge_id: number;
       }
     ): Promise<boolean> => {
-      return badgeService.deleteCustomBadge(badge_id);
+      const badge = await prisma.badge.findUnique({
+        where: { badge_id },
+      });
+
+      if (!badge) {
+        throw new Error(`Badge with ID ${badge_id} does not exist`);
+      }
+
+      if (badge.badge_type !== BadgeType.CUSTOM) {
+        throw new Error(
+          `Badge with ID ${badge_id} is not a custom badge and cannot be deleted`
+        );
+      }
+
+      await prisma.badge.delete({
+        where: { badge_id },
+      });
+
+      return true;
     },
     editBadgeLevel: async (
       _parent: undefined,
@@ -134,12 +284,19 @@ const badgeResolver = {
         marillac_bucks: number;
       }
     ): Promise<boolean> => {
-      return badgeService.editBadgeLevel(
-        badge_id,
-        badge_level,
-        benchmark,
-        marillac_bucks
-      );
+      await prisma.badgeLevel.update({
+        where: {
+          badge_id_level: {
+            badge_id,
+            level: badge_level,
+          },
+        },
+        data: {
+          benchmark,
+          marillac_bucks,
+        },
+      });
+      return true;
     },
   },
 };
